feat: add e2e test
