+++ src/tracing/utils.ts
@@ -1,6 +1,5 @@
 /**
- * Returns the time elapsed between page load and now,
- * if the time-to-swap mark doesn't already exist.
+ * Returns the time elapsed between page load and now.
  */
 export function calculateElapsedTimeWithPerformanceMark(markName: string): number {
   const elapsedTime = performance.mark(markName)
